+++ packages/eslint-config/shared/typescript-react.js
@@ -1,0 +1,6 @@
+module.exports = {
+  rules: {
+    'react/jsx-filename-extension': ['error', { extensions: ['.js', '.tsx'] }],
+    'react/require-default-props': 'off',
+  },
+};